Weather_app/requirements.txt
1.flask
2.gunicorn
3.requests